CNC_release/requirements.txt
charset-normalizer==3.1.0
cramjam==2.7.0
cycler==0.11.0
fastparquet==0.8.1
fonttools==4.38.0
fsspec==2023.1.0
idna==3.4
imageio==2.31.1
kiwisolver==1.4.4
kornia==0.6.12
lpips==0.1.4
markdown-it-py==2.2.0
mdurl==0.1.2
nerfacc==0.5.3
numpy==1.21.6
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.8.1.78
packaging==23.1
Pillow==9.5.0
pyarrow==12.0.1
Pygments==2.15.1
pyinstrument==4.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rich==13.4.2
six==1.16.0
torch==1.13.1
torchac==0.9.3
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
